Unicode sandwich calls to dmutils.email.helpers.hash_string
