+++ dmutils/__init__.py
@@ -1,4 +1,4 @@
-from . import logging, config, proxy_fix, formats, request_id
+from . import config, proxy_fix, formats, request_id
 from .flask_init import init_app, init_manager
 
 import flask_featureflags
